Move body generator for enums to its own module
Cut-paste
